requirements.txt
pytest
pytest-html
pytest-cov
selenium
pymysql
flask
py
webdriver-manager